Moved scenario projections to orchastrators and updated docstring
